% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IsoCorrectoR}
\alias{IsoCorrectoR}
\title{IsoCorrectoR example data}
\format{A list with four elements, namely
\describe{
\item{tool_features: }{overview of correction features as provided by various tools IsoCorrectoR is compared with.}
\item{element_file: }{example data for an Element file. This information is independent of the resolution type (normal or high)}
\item{normal_resolution: }{list with two elements, containing examples for normal resolution data: molecule information and measurement data}
\item{high_resolution: }{list with two elements, containing examples for high resolution data: molecule information and measurement data}
}}
\usage{
data(IsoCorrectoR)
}
\description{
Example data for the IsoCorrectoR package
}
\author{
Paul Heinrich, Christian Kohler
}
\keyword{datasets}
